+++ src/DenseInverse.cpp
@@ -11,7 +11,7 @@ void DenseInverse<real>::backSubstituteUpperTriangular(size_t m, size_t n, real*
 	assert(m >= n);
 	for (int i = n-1; i >= 0; --i) {
 		real sum = 0;
-		for (int j = i+1; j < n; ++j) {
+		for (int j = i+1; j < (int)n; ++j) {
 			sum += a[i+m*j] * x[j];
 		}
 		x[i] = (b[i] - sum) / a[i+m*i];
@@ -22,10 +22,10 @@ template<typename real>
 void HouseholderQR<real>::applyQ(real* a, int m, int k, int jmin, int jmax, real* v) {
 	for (int j = jmin; j < jmax; ++j) {
 		real vDotMj = 0;
-		for (int i = k; i < m; ++i) {
+		for (int i = k; i < (int)m; ++i) {
 			vDotMj += v[i-k] * a[i + m * j];
 		}
-		for (int i = k; i < m; ++i) {
+		for (int i = k; i < (int)m; ++i) {
 			a[i + m * j] -= 2. * vDotMj * v[i-k];
 		}
 	}
@@ -37,20 +37,20 @@ void HouseholderQR<real>::householderQR(size_t m, size_t n, real* qt, real* a) {
 
 	real v[m];
 
-	for (int i = 0; i < m; ++i) {
-		for (int j = 0; j < m; ++j) {
+	for (int i = 0; i < (int)m; ++i) {
+		for (int j = 0; j < (int)m; ++j) {
 			qt[i+m*j] = i == j ? 1 : 0;
 		}
 	}
 
-	for (int k = 0; k < n; ++k) {
+	for (int k = 0; k < (int)n; ++k) {
 		//v[i-k] = a[i,k], k<=i<m
 		memcpy(v, a + k + m * k, sizeof(real) * (m - k));
 		real vLen = Vector<real>::normL2(m-k, v);
 		v[0] += vLen * (v[0] < 0 ? -1 : 1);
 		vLen = Vector<real>::normL2(m-k, v);
 		if (vLen > 1e-10) {
-			for (int i = 0; i < m-k; ++i) {
+			for (int i = 0; i < (int)m-k; ++i) {
 				v[i] /= vLen;
 			}
 		}
@@ -71,9 +71,9 @@ void HouseholderQR<real>::solveLinear_leastSquares(size_t m, size_t n, real* x, 
 	//if x == b are matching pointers then I wouldn't be able to write qtb to x without corrupting it
 	//so I'll use an extra buffer to store the intermediate value
 	real qtb[m];
-	for (int i = 0; i < m; ++i) {
+	for (int i = 0; i < (int)m; ++i) {
 		real sum = 0;
-		for (int j = 0; j < m; ++j) {
+		for (int j = 0; j < (int)m; ++j) {
 			sum += qt[i + m * j] * b[j];
 		}
 		qtb[i] = sum;
@@ -96,19 +96,19 @@ void HouseholderQR<real>::matrixInverse(size_t n, real* ainv, const real* a) {
 		
 	real qty[n];
 
-	for (int i = 0; i < n; ++i) {
-		for (int j = 0; j < n; ++j) {
+	for (int i = 0; i < (int)n; ++i) {
+		for (int j = 0; j < (int)n; ++j) {
 			ainv[i+n*j] = i == j ? 1 : 0;
 		}
 	}
-	for (int j = 0; j < n; ++j) {
+	for (int j = 0; j < (int)n; ++j) {
 		//solve for x in a x = y
 		//let a = q r
 		//q r x = y
 		//r x = q^t y
-		for (int i = 0; i < n; ++i) {
+		for (int i = 0; i < (int)n; ++i) {
 			real sum = 0;
-			for (int k = 0; k < n; ++k) {
+			for (int k = 0; k < (int)n; ++k) {
 				sum = sum + qt[i + n * k] * ainv[k + n * j];
 			}
 			qty[i] = sum;
